Small change
Calculate n as np.int64(mean + sigma * 2), as n must be an integer

+++ preliz/distributions/discrete.py
@@ -193,11 +193,11 @@ class BetaBinomial(Discrete):
         # Crude aproximation for n (as in Binomial distribution)
         # For alpha and beta see:
         # https://en.wikipedia.org/wiki/Beta-binomial_distribution#Method_of_moments
-        n = mean + sigma * 2
+        n = np.int64(mean + sigma * 2)
         p = mean / n
         rho = ((sigma**2 / (mean * (1 - p))) - 1) / (n - 1)
         alpha = max(0.5, (p / rho) - p)
-        beta =  max(0.5, (alpha / p) - alpha)
+        beta = max(0.5, (alpha / p) - alpha)
         self._update(alpha, beta, n)
 
     def _fit_mle(self, sample):
